pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coloredcurves4yolo"
version = "0.1.0"
description = "Export YOLO validation curve data and colorful grouped plots."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Repository Contributors" }
]
classifiers = [
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "matplotlib>=3.7",
  "numpy>=1.23",
  "PyYAML>=6.0",
  "ultralytics>=8.0.0",
]

[project.scripts]
coloredcurves4yolo = "coloredcurves4yolo.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
